route53recoveryreadiness: Use service package 'withExtraOptions' mechanism.
