+++ internal/service/route53recoveryreadiness/service_package_gen.go
@@ -5,6 +5,10 @@ package route53recoveryreadiness
 import (
 	"context"
 
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/route53recoveryreadiness"
+	"github.com/hashicorp/aws-sdk-go-base/v2/endpoints"
+	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/hashicorp/terraform-provider-aws/internal/conns"
 	"github.com/hashicorp/terraform-provider-aws/internal/types"
 	"github.com/hashicorp/terraform-provider-aws/names"
@@ -63,6 +67,48 @@ func (p *servicePackage) SDKResources(ctx context.Context) []*types.ServicePacka
 
 func (p *servicePackage) ServicePackageName() string {
 	return names.Route53RecoveryReadiness
+}
+
+// NewClient returns a new AWS SDK for Go v2 client for this service package's AWS API.
+func (p *servicePackage) NewClient(ctx context.Context, config map[string]any) (*route53recoveryreadiness.Client, error) {
+	cfg := *(config["aws_sdkv2_config"].(*aws.Config))
+	optFns := []func(*route53recoveryreadiness.Options){
+		route53recoveryreadiness.WithEndpointResolverV2(newEndpointResolverV2()),
+		withBaseEndpoint(config[names.AttrEndpoint].(string)),
+		func(o *route53recoveryreadiness.Options) {
+			switch partition := config["partition"].(string); partition {
+			case endpoints.AwsPartitionID:
+				if region := endpoints.UsWest2RegionID; cfg.Region != region {
+					tflog.Info(ctx, "overriding region", map[string]any{
+						"original_region": cfg.Region,
+						"override_region": region,
+					})
+					o.Region = region
+				}
+			}
+		},
+		withExtraOptions(ctx, p, config),
+	}
+
+	return route53recoveryreadiness.NewFromConfig(cfg, optFns...), nil
+}
+
+// withExtraOptions returns a functional option that allows this service package to specify extra API client options.
+// This option is always called after any generated options.
+func withExtraOptions(ctx context.Context, sp conns.ServicePackage, config map[string]any) func(*route53recoveryreadiness.Options) {
+	if v, ok := sp.(interface {
+		withExtraOptions(context.Context, map[string]any) []func(*route53recoveryreadiness.Options)
+	}); ok {
+		optFns := v.withExtraOptions(ctx, config)
+
+		return func(o *route53recoveryreadiness.Options) {
+			for _, optFn := range optFns {
+				optFn(o)
+			}
+		}
+	}
+
+	return func(*route53recoveryreadiness.Options) {}
 }
 
 func ServicePackage(ctx context.Context) conns.ServicePackage {
